token authorization added to own endpoints

+++ stellar/stellar_backend/controllers/messages_controller.py
@@ -1,13 +1,22 @@
-from django.shortcuts import render
 from rest_framework.decorators import api_view
-from ..services import message_service
+from rest_framework import status
+from rest_framework.response import Response
+
+from ..services import message_service, session_service
+from ..static import constants as const
 
 @api_view(['POST'])
 def new_message(request):
+    if not session_service.is_token_valid(request.META):
+        return Response(const.INVALID_TOKEN, status = status.HTTP_401_UNAUTHORIZED)
+    
     return message_service.new_message(request)
 
 @api_view(['GET'])
 def get_message_content(request, message_id):
+    if not session_service.is_token_valid(request.META):
+        return Response(const.INVALID_TOKEN, status = status.HTTP_401_UNAUTHORIZED)
+    
     return message_service.get_message_content(message_id)
 
 @api_view(['GET'])
@@ -16,6 +25,9 @@ def get_group_messages(request, group_id):
 
 @api_view(['DELETE'])
 def delete_message(request, message_id):
+    if not session_service.is_token_valid(request.META):
+        return Response(const.INVALID_TOKEN, status = status.HTTP_401_UNAUTHORIZED)
+    
     return message_service.delete_message(message_id)
 
 @api_view(['DELETE'])
